refactor: :recycle: Correct admin dashboard authentication
