+++ .gitignore
@@ -29,6 +29,7 @@ yarn-error.log*
 # local env files
 .env*.local
 .env
+.env.development
 
 # vercel
 .vercel
