+++ pages/admin/permission.tsx
@@ -1,5 +1,5 @@
+import { AdminPageLayout } from '@/module/layout/AdminLayout';
 import React from 'react';
-import { AdminPageLayout } from '.';
 
 export default function Permission() {
   return <div>In progress permission</div>;
