+++ pages/admin/users.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { AdminPageLayout } from '.';
 import { UserType } from '@/utils/types/user';
 import { findUsers } from '@/apps/api/modules/user/user.utils';
 import { getAllUsersServices } from '@/apps/api/modules/user/user.services';
 import { useQuery } from '@tanstack/react-query';
 import { UserTable } from '@/module/admin/components/userTable/UserTable';
 import { Box } from '@chakra-ui/react';
+import { AdminPageLayout } from '@/module/layout/AdminLayout';
 
 interface AdminUserProps {
   users: UserType[];
